Add config to disable automatic HTTP redirects
Previously there was no way to configure whether the HTTP client
should follow HTTP redirects automatically or not. This commit adds
an environment variable to allow configuring it on a per-run basis.

Setting "MAESTRO_FOLLOW_REDIRECTS" to "false" will disable redirects.
Setting it to "true" or leaving it unsetted will default to follow
redirects as before.