Error message improvements (#1372)
Error message improvements

* Proper error messages for empty workspaces
* Surface workspace validation errors before connecting to device when running `maestro test`
* Lay test fixture groundwork for future error improvements

Behavior Change:
This PR also changes the behavior when includeTags are specified in both config.yaml and as a command line option. Previously, in order for a flow to be included, it would need to match one of the config.yaml includeTags AND one of the command line includeTags. Even though this might be useful for some specific scenarios, this behavior is confusing and hard to explain. Now, in order for a flow to be included, it needs to match only one includeTag regardless of where the includeTag was specified.