feat: add optional label attribute for all Maestro commands for custom descriptions  (#1292)
Adds a label attribute to all commands. This adds readability to the flow file as to give a better description of a particular step (especially relevant for tapping on points or vague element selectors) and is also returned in the description() method of those commands to be displayed in the CLI output. This also solves secrets being printed from the CLI output.

---------

Co-authored-by: Brett James <super_cockroach@yahoo.com>
Co-authored-by: igorsmotto <igor@mobile.dev>