feat (maestro): add originalDescription property to commands
Context
The goal of this PR is to expose the properties of the original description of the file - the one that would exist without the label.

What do I mean by that?

When we have a command, we have the properties:

label
description()
The description represents the human-readable form of the command. Each command has a corresponding "rule" or formula to create such a format. However, when the command has a label, this formula is not accessible: we always return the label in the description() function when it is present.

So the goal is to allow us to see the original Description (the one without the label)

High-level reason for why do so
-> so we can store, in Maestro Cloud (and possibly other scenarios) both the labeled description and the unlabeled one, allowing us to choose which one to see.

Proposed changes
Create the property originalDescription: exactly the same as description(), except it ignores the label
Make the description be a choice between label and originalDescription: if label is present, return label

Testing
Wrote test cases to confirm the correct accessibility of the properties.