Allow forwarding to fail to allow studio and flow to be active at the same time
Fixes #2188

This is a bit of a workaround, both studio and flow forwards the
driver-host-port (default 7001) for the android driver session.  The
forwarding fails for the second instance as the port is already open.
But it doesn't really matter which driver-host is used, so let
forwarding silently fail and let the android driver session continue.