Better detection of web flows
We've used a string regex to determine if a flow is a web flow, but this
fails when flows use JS expressions. Instead, we're preserving
the `url:` field in the YamlConfig, but then mapping it to appId in
the MaestroConfig. This allows the YamlConfig to more accurately reflect
what's in the raw yaml, and allows us to just use the presence of the
`url:` field to determine if a flow is a web flow.