Better detection of web flows (#2611)
This commit now requires flows to use the `url` field in the Yaml config if they want to 
use a URL; they cannot pass a URL in the `appId` field.

The rationale: in multiple areas in our codebase (cli `test`, cli `cloud`. `studio-server`) we need to know
if a flow is a web flow. We use `FileUtils.isWebFlow` to determine this based on a regex
on the `appId` field. This is a problem for flows that use JS scripts in the appId field, since the
"https://" won't be populated at that point. 

Instead of trying to instantiate a JS engine, we're going to require flows to use the `url` field to 
specify the web flow. This has other benefits in making it clear if a flow targets a browser or a device.